pyproject.toml
[project]
name = "YNAmazon"
dynamic = ["version"]
description = "YNAmazon - A program to annotate YNAB transactions with Amazon order info "
readme = "README.md"
requires-python = ">=3.9,<3.13"
license = "MIT"
dependencies = [
    "amazon-orders>=3.2.15",
    "loguru>=0.7.3",
    "pydantic[email]>=2.11.2",
    "pydantic-settings>=2.8.1",
    "rich>=14.0.0",
    "ynab>=1.3.1",
    "typer>=0.15.2",
    "pydantic-core>=2.33.1",
    "furl>=2.1.4",
    "requests>=2.30.0",
    "toml>=0.10.2",
    "pyyaml>=6.0.2",
]

[project.urls]
Repository = "https://github.com/DanielKarp/YNAmazon"
Issues = "https://github.com/DanielKarp/YNAmazon/issues"

[tool.hatch.version]
path = "src/ynamazon/__init__.py"

[project.scripts]
yna = "ynamazon.cli.cli:cli"

[dependency-groups]
dev = [
    "deptry>=0.23.0",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "ruff>=0.11.4",
    "types-pyyaml>=6.0.12.20250402",
    "types-requests>=2.32.0.20250328",
    "types-toml>=0.10.8.20240310",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[[tool.mypy.overrides]]
module= ["amazonorders.*"]
ignore_missing_imports = true

# [tool.deptry]
# known_first_party = ["ynamazon"]
